+++ fabric/src/main/java/net/dodogang/plume/ash/registry/fabric/FuelRegistryImpl.java
@@ -1,34 +1,42 @@
 package net.dodogang.plume.ash.registry.fabric;
 
 import net.dodogang.plume.ash.registry.FuelRegistry;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemConvertible;
+import net.minecraft.tag.Tag;
 import org.jetbrains.annotations.ApiStatus;
 
 @ApiStatus.Internal
 public final class FuelRegistryImpl {
     /**
-     * Registers an item as fuel with a burn time for furnace-like blocks.
+     * Registers items as fuel with a burn time for furnace-like blocks.
      *
-     * <p>Note: Undefined behavior if the item has been registered already by
-     * vanilla or another mod. Expect this to not affect those items.
+     * <p>If an item has already been registered a burn time by vanilla or
+     * another mod, this doesn't overwrite it.
      *
-     * @param item an item
      * @param burnTime the item's burn time
+     * @param items items
      */
-    public static void register(ItemConvertible item, int burnTime) {
-        net.fabricmc.fabric.api.registry.FuelRegistry.INSTANCE.add(item, burnTime);
+    public static void register(int burnTime, ItemConvertible... items) {
+        for (ItemConvertible item : items) {
+            if (net.fabricmc.fabric.api.registry.FuelRegistry.INSTANCE.get(item) < 0) {
+                net.fabricmc.fabric.api.registry.FuelRegistry.INSTANCE.add(item, burnTime);
+            }
+        }
     }
 
     /**
-     * Unregisters an item as fuel furnace-like blocks.
+     * Registers items from a tag as fuel with a burn time for furnace-like
+     * blocks.
      *
-     * <p>Note: This only unregisters it if it has been registered through
-     * plume's {@link FuelRegistry}.
+     * <p>If an item has already been registered a burn time by vanilla or
+     * another mod, this doesn't overwrite it.
      *
-     * @param item an item
+     * @param burnTime the item's burn time
+     * @param itemTag an item tag
      */
-    public static void unregister(ItemConvertible item) {
-        net.fabricmc.fabric.api.registry.FuelRegistry.INSTANCE.remove(item);
+    public static void registerTag(int burnTime, Tag<Item> itemTag) {
+        net.fabricmc.fabric.api.registry.FuelRegistry.INSTANCE.add(itemTag, burnTime);
     }
 
     /**